src/requirements.txt
beautifulsoup4==4.11.1
dateparser==1.1.4
Django==4.1.3
django-environ==0.9.0
djangorestframework==3.14.0
django-cors-headers==3.13.0
elasticsearch==7.17.8
elasticsearch-dsl==7.4.0
feedparser==6.0.10
future==0.18.2
gunicorn==20.1.0
huggingface-hub==0.11.1
json5==0.9.10
langcodes==3.3.0
Levenshtein==0.20.9
lxml==4.9.2
Markdown==3.4.1
MarkupSafe==2.1.1
nltk==3.7
numpy==1.23.5
opendatasets==0.1.22
protobuf==3.20.1
regex==2022.10.31
requests==2.28.1
sentence-transformers==2.2.2